[14.0][IMP] product_multi_company: Add tests common class
In order to extend testing, a common class is necessary to avoid code
duplication.
