+++ product_multi_company/tests/__init__.py
@@ -1,1 +1,2 @@
+from . import common
 from . import test_product_multi_company
